Pipe update

+++ Resonator/Resonator-Pipe.hpp
@@ -29,7 +29,7 @@ public:
 		}())
 	, recur(0.999f)
 	, delay([&] {
-			const double length = (static_cast<double>(sampleRate)/frequency) - 1.0 - (static_cast<double>(convolution.size()) / 2.0);
+			const double length = (sampleRate/frequency) - 1.0 - (static_cast<double>(convolution.size()) / 2.0);
 			return Delay::Lagrange<3>(std::ceil(length), length);
 		}())
 	{}
